kafka/consumers/retry_consumer/requirements.txt
confluent-kafka==2.12.2
python-dotenv==1.2.1